+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+* Add placeholder option to Select (GUIC-205)
+
 ### Fixes
 
 * Fix the order of ngModel and onChange (GUIC-204)